+++ src/SpongebobRevengeOfTheFlyingDutchman/bitmap.ts
