+++ src/SpongebobRevengeOfTheFlyingDutchman/hfog.ts
