+++ src/SpongebobRevengeOfTheFlyingDutchman/light.ts
