+++ src/SpongebobRevengeOfTheFlyingDutchman/materialanim.ts
