+++ src/SpongebobRevengeOfTheFlyingDutchman/mesh.ts
