+++ src/SpongebobRevengeOfTheFlyingDutchman/node.ts
