+++ src/SpongebobRevengeOfTheFlyingDutchman/omni.ts
