+++ src/SpongebobRevengeOfTheFlyingDutchman/render.ts
