+++ src/SpongebobRevengeOfTheFlyingDutchman/scenes.ts
