+++ src/SpongebobRevengeOfTheFlyingDutchman/skin.ts
