+++ src/SpongebobRevengeOfTheFlyingDutchman/util.ts
